Use fuzzy matching for workspace symbols
