+++ packages/langium/src/services.ts
@@ -58,6 +58,7 @@ import type { DocumentationProvider } from './documentation/documentation-provid
 import type { InlayHintProvider } from './lsp/inlay-hint-provider';
 import type { WorkspaceSymbolProvider } from './lsp/workspace-symbol-provider';
 import type { NodeKindProvider } from './lsp/node-kind-provider';
+import type { FuzzyMatcher } from './lsp/fuzzy-matcher';
 
 /**
  * The services generated by `langium-cli` for a single language. These are derived from the
@@ -158,6 +159,7 @@ export type LangiumDefaultSharedServices = {
         InlayHintProvider?: InlayHintProvider
         WorkspaceSymbolProvider?: WorkspaceSymbolProvider
         NodeKindProvider: NodeKindProvider
+        FuzzyMatcher: FuzzyMatcher
         LanguageServer: LanguageServer
     }
     workspace: {
